Use a RecyclerView with Firebase to display a scrolling list of data
- Replace the TextView (for testing if data comes along with the eventListener) and put in place a RecyclerView.
A RecyclerView allows scrolling list of data.
Each row of a RecyclerView contains a TextView (each entry is one element). For this we add a new element a TextView to show the title (LinearLayout)

In order for the RecyclerView to work, dus, get data, it needs an adapter.
The adapter sends data to the RecyclerView through an object called ViewHolder. The ViewHolder describes an item of the RecyclerView. In this app, it is a single row of data containing the travel deal information (it also contains info about its place in the RecyclerView)
The adapter should subclass the ViewHolder to load the views, and then bind the data to the views.
ViewHolders get cached to make scrolling smoother, because the findViewById method gets called only when an item is created, and NOT each time it's updated.

Finally, we have the LayoutManager. We can use this one in order to decide how the items are displayed in a RecyclerView.